Move confirmation_prompt to discord_ui module
Renamed confirmation_prompt.py from shared/ to shared/discord_ui/ for better organization. Updated import in music_command_cog.py to reflect new location.
